refactor: 디렉터리 별 index.js 생성으로 모듈 경로 정리
