+++ src/utils/index.js
@@ -1,0 +1,11 @@
+const Console = require('./Console');
+const inputErrorHandler = require('./inputErrorHandler');
+const randomNumberGenerator = require('./randomNumberGenerator');
+const scoreConverter = require('./scoreConverter');
+
+module.exports = {
+  Console,
+  inputErrorHandler,
+  randomNumberGenerator,
+  scoreConverter,
+};
